+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.0.0-beta5] - In progress
+
+### Added
+
+- New method `Collection::updateHook(): void` to manipulate the collection after an update
+
 ## [1.0.0-beta4] - 2021-09-20
 
 ### Added